+++ packages/core/styles/src/properties/index.ts
@@ -1,2 +1,3 @@
 export * from "./css";
 export * from "./base";
+export * from "./typography";